requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.31.1
numpy==1.26.4
pillow==10.4.0
keras==3.4.1
tensorflow==2.17.0
python-multipart==0.0.9
requests==2.32.3
opencv-python-headless==4.10.0.84
streamlit==1.39.0
